Add refund sync and refactor payment/email flow

+++ ToDoApi/Model/Dto/InternalFormListDto.cs
@@ -7,7 +7,7 @@
         public string OrderType { get; set; }
         public DateOnly OrderDate { get; set; }
         public bool Paid { get; set; }
-        public float? AmountPaid { get; set; }
+        public decimal? AmountPaid { get; set; }
         public string Address { get; set; }
         public string ContactNo { get; set; }
         public string OrderNo { get; set; }
