+++ ToDoApi/Model/Dto/OrderDto.cs
@@ -26,12 +26,12 @@
         public string? Note { get; set; }
         public string? Changes { get; set; }
         public string? PackNote { get; set; }
-        public float? AmountPaid { get; set; }
-        public float? AmountOwning { get; set; }
-        public float? Donation { get; set; }
-        public float? Freight { get; set; }
-        public float? Refund { get; set; }
-        public float? AdminCharges { get; set; }
+        public decimal? AmountPaid { get; set; }
+        public decimal? AmountOwning { get; set; }
+        public decimal? Donation { get; set; }
+        public decimal? Freight { get; set; }
+        public decimal? Refund { get; set; }
+        public decimal? AdminCharges { get; set; }
         public DateOnly? PayBy { get; set; }
         public int? Status { get; set; }
     }
